requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
setuptools>=58.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0

